# Use lookup table instead of switch for selecting byte parser.