While Go may compile a packed switch into a lookup table, it doesn't seem to here maybe because it's too many cases, or because of the non-singular case statements. 

Just small in the large scope of things but removed some noise from the cpu profile

After
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/libinjection-go
BenchmarkSQLiDriver
BenchmarkSQLiDriver/sqli
BenchmarkSQLiDriver/sqli-10         	   20707	     56244 ns/op
BenchmarkSQLiDriver/folding
BenchmarkSQLiDriver/folding-10      	    9721	    118566 ns/op
BenchmarkSQLiDriver/tokens
BenchmarkSQLiDriver/tokens-10       	    5677	    204501 ns/op
PASS
```

Before
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/libinjection-go
BenchmarkSQLiDriver
BenchmarkSQLiDriver/sqli
BenchmarkSQLiDriver/sqli-10         	   19716	     57947 ns/op
BenchmarkSQLiDriver/folding
BenchmarkSQLiDriver/folding-10      	    9536	    118126 ns/op
BenchmarkSQLiDriver/tokens
BenchmarkSQLiDriver/tokens-10       	    5713	    206581 ns/op
PASS
```