@jcchavezs 2022-09-08
It looks good to me. My only question would be if itsn't better to generate such map with a go script and run the `go generate` to have this statically.