@jptosso 2022-09-08
I agree with @jcchavezs but we should handle that in other issue

Also we could run some automatic optimization, like:
```go
switch 
case parse >= 0 && parse < 33:
  return parseWhite(s)
```